feat(TUP-29014):backport 2021-03-08 (#5918)
* feat(TUP-29014):backport 2021-03-08

* feat(TUP-29014):bugfix 2021-03-09

+++ main/plugins/org.talend.designer.runprocess/src/main/java/org/talend/designer/runprocess/java/TalendJavaProjectManager.java
@@ -422,10 +422,10 @@ public class TalendJavaProjectManager {
         talendCodesJarJavaProjects.remove(getCodesJarProjectId(info));
     }
 
-    public static void deleteTalendCodesJarProject(Property property, boolean deleteContent) {
+    public static void deleteTalendCodesJarProject(CodesJarInfo info, boolean deleteContent) {
+        Property property = info.getProperty();
         ERepositoryObjectType type = ERepositoryObjectType.getItemType(property.getItem());
-        String projectTechName = ProjectManager.getInstance().getProject(property).getTechnicalLabel();
-        deleteTalendCodesJarProject(type, projectTechName, property.getLabel(), deleteContent);
+        deleteTalendCodesJarProject(type, info.getProjectTechName(), property.getLabel(), deleteContent);
     }
 
     public static void deleteTalendCodesJarProject(ERepositoryObjectType type, String projectTechName, String codesJarName,
@@ -433,15 +433,16 @@ public class TalendJavaProjectManager {
         try {
             String projectId = getCodesJarProjectId(type, projectTechName, codesJarName);
             ITalendProcessJavaProject project = talendCodesJarJavaProjects.get(projectId);
-            if (project != null) {
-                IFolder folder = new AggregatorPomsHelper().getCodeFolder(type).getFolder(codesJarName);
-                IFile projectPom = folder.getFile(TalendMavenConstants.POM_FILE_NAME);
-                AggregatorPomsHelper.removeFromParentModules(projectPom);
+
+            if (project != null && project.exists()) {
                 project.getProject().delete(deleteContent, true, null);
             }
             if (deleteContent) {
                 IFolder folder = new AggregatorPomsHelper(projectTechName).getCodeFolder(type).getFolder(codesJarName);
-                folder.delete(true, false, null);
+                AggregatorPomsHelper.removeFromParentModules(folder.getFile(TalendMavenConstants.POM_FILE_NAME));
+                if (folder.exists()) {
+                    folder.delete(true, false, null);
+                }
             }
             talendCodesJarJavaProjects.remove(projectId);
         } catch (Exception e) {
@@ -460,7 +461,7 @@ public class TalendJavaProjectManager {
                 Property property = project.getPropery();
                 if (property != null) {
                     IPath jobPath = ItemResourceUtil.getItemRelativePath(property);
-                    if (folderPath.isPrefixOf(jobPath)) {
+                    if (folderPath.isPrefixOf(jobPath) && project.exists()) {
                         project.getProject().delete(deleteContent, true, null);
                         iterator.remove();
                     }
@@ -470,7 +471,9 @@ public class TalendJavaProjectManager {
                 // delete folder
                 AggregatorPomsHelper helper = new AggregatorPomsHelper();
                 IFolder folder = helper.getProcessFolder(processType).getFolder(folderPath);
-                folder.delete(true, false, null);
+                if (folder.exists()) {
+                    folder.delete(true, false, null);
+                }
             }
         } catch (CoreException e) {
             ExceptionHandler.process(e);
@@ -509,7 +512,9 @@ public class TalendJavaProjectManager {
             String key = iterator.next();
             if (key.contains(id)) {
                 ITalendProcessJavaProject projectToDelete = talendJobJavaProjects.get(key);
-                projectToDelete.getProject().delete(deleteContent, true, null);
+                if (projectToDelete.exists()) {
+                    projectToDelete.getProject().delete(deleteContent, true, null);
+                }
                 String version = key.split("\\|")[1]; //$NON-NLS-1$
                 removedVersions.add(version);
                 iterator.remove();
